pyproject.toml
[project]
name = "dia-test-harness"
version = "0.1.0"
description = "Test harness for Data Insights Agent (DIA)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-cloud-bigquery>=3.13.0",
    "faker>=19.3.0",
    "pandas>=2.1.0",
    "click>=8.1.7",
    "db-dtypes>=1.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
    "google-auth>=2.23.0",
    "google-cloud-aiplatform>=1.38.0",
    "openpyxl>=3.1.0",
]

[project.scripts]
dia-harness = "orchestrator.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/orchestrator", "src/evaluation", "src/iterative"]

[tool.pytest.ini_options]
testpaths = ["tests"]
